backend/python_main/requirements.txt
mysql-connector-python
numpy
pandas
polars
schedule
redis
python-dotenv